Retraction Bugfix
Added a failsafe for retraction of the hook. THe player controller now has information on whether or not the hook is retracted, and sets a boolean on the HookController to indicate if it triggered a launch. Only the HookController can revert that boolean though, through actually retracting the hook.

Also made it so that if the hook is out, and the player is not pressing any mouse buttons, it gets retracted.

This should prevent the weird unsyncing of the mouse buttons with the state of the grappling hook.